Choose field layout algorithm via TypeSystemContext
 - Make the field layout algorithm selectable by the TypeSystemContext
 - Move ContainsPointer computation to field layout algorithm
